🦄 refactor(layout): extract layout from default layout
